+++ src/Sql/Sql/ChangeLog.md
@@ -18,6 +18,14 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Add support for SQL DB Hyperscale tier
+
+## Version 1.2.0
+* Add Get/Set AzSqlDatabaseBackupShortTermRetentionPolicy
+* Fix issue where not being logged into Azure account would result in nullref exception when executing SQL cmdlets
+* Fixed null ref exception in Get-AzSqlCapability
+
+## Version 1.1.0
 * Update incorrect online help URLs
 * Updated parameter description for LicenseType parameter with possible values
 * Fix for updating managed instance identity not working when it is the only updated property